move create-git-instance to utils folder
