+++ src/utils/create-git-instance.ts
@@ -1,7 +1,10 @@
 import path from "path";
 import simpleGit from "simple-git/promise";
-import { findRepoRoot } from "./utils/find-repo-root";
+import { findRepoRoot } from "./find-repo-root";
 
+/**
+ * Creates a `SimpleGit` instance that uses the git repository containing the given `packageJsonPath`
+ */
 export async function createGitInstance(packageJsonPath: path.ParsedPath) {
   const gitRepoPath = await findRepoRoot(packageJsonPath);
   const gitRepoPathStr = path.format(gitRepoPath);
